Move Menu to widgets
