+++ src/widgets/menu/ui/entry/index.tsx
